Slightly improve the "trim trailing space" behavior
Right before saving to disk, we invoke the
`Document.trim_trailing_spaces` method, which will rewrite all rows by
trimming all trailing spaces at the end.

We also move the cursor to the last character of the row if it happened
to be located on one of the trailing spaces that just got trimmed.